requirements.txt
streamlit==1.22.0
snowflake-snowpark-python==0.7.0
pyarrow==10.0.1
pandas==1.4.2
numpy==1.21.5
unstructured==0.6.6
tiktoken==0.4.0